move interface definitions into interfaces.ts
